Atualizado feature com a branch da estória.
