+++ teste/SME.SGP.TesteIntegracao/Listao/Ao_alterar_frequencia.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using MediatR;
@@ -108,8 +107,7 @@ namespace SME.SGP.TesteIntegracao.Listao
             registroFrequenciaId.ShouldBeGreaterThan(0);
             
             var frequenciasSalvar = listaAulaId.Select(aulaId => new FrequenciaSalvarAulaAlunosDto
-                {FrequenciaId = registroFrequenciaId,AulaId = aulaId, Alunos = ObterListaFrequenciaSalvarAluno() }).ToList();
-            
+                {FrequenciaId = registroFrequenciaId,AulaId = aulaId, Alunos = ObterListaFrequenciaSalvarAlunoComAusencia() }).ToList();
             
             var useCaseSalvar = InserirFrequenciaListaoUseCase();
             useCaseSalvar.ShouldNotBeNull();
@@ -118,26 +116,6 @@ namespace SME.SGP.TesteIntegracao.Listao
             retornoSalvar.ShouldNotBeNull();
 
             retornoSalvar.AlteradoEm!.Value.Date.ShouldBeEquivalentTo(DateTime.Now.Date);
-        }
-        private IEnumerable<FrequenciaSalvarAlunoDto> ObterListaFrequenciaSalvarAluno()
-        {
-            return CODIGOS_ALUNOS.Select(codigoAluno => new FrequenciaSalvarAlunoDto
-                { CodigoAluno = codigoAluno, Frequencias = ObterFrequenciaAula(codigoAluno) }).ToList();
-        }
-        private IEnumerable<FrequenciaAulaDto> ObterFrequenciaAula(string codigoAluno)
-        {
-            string[] codigosAlunosAusencia = { CODIGO_ALUNO_1, CODIGO_ALUNO_3 };
-            string[] codigosAlunosPresenca = { CODIGO_ALUNO_2, CODIGO_ALUNO_4, CODIGO_ALUNO_6 };
-            string[] codigosAlunosRemotos = { CODIGO_ALUNO_5 };
-
-            return QUANTIDADES_AULAS.Select(numeroAula => new FrequenciaAulaDto
-            {
-                NumeroAula = numeroAula,
-                TipoFrequencia = codigosAlunosAusencia.Contains(codigoAluno) ? TipoFrequencia.F.ObterNomeCurto() :
-                    codigosAlunosPresenca.Contains(codigoAluno) ? TipoFrequencia.C.ObterNomeCurto() :
-                    codigosAlunosRemotos.Contains(codigoAluno) ? TipoFrequencia.R.ObterNomeCurto() :
-                    TIPOS_FREQUENCIAS[new Random().Next(TIPOS_FREQUENCIAS.Length)].ObterNomeCurto()
-            }).ToList();
         }
     }
 }
