+++ teste/SME.SGP.TesteIntegracao/Listao/Base/ListaoTesteBase.cs
@@ -88,17 +88,39 @@ namespace SME.SGP.TesteIntegracao.Listao
         protected IEnumerable<FrequenciaSalvarAlunoDto> ObterListaFrequenciaSalvarAluno()
         {
             return CODIGOS_ALUNOS.Select(codigoAluno => new FrequenciaSalvarAlunoDto
-                { CodigoAluno = codigoAluno, Frequencias = ObterFrequenciaAula(codigoAluno) }).ToList();
+                { CodigoAluno = codigoAluno, Frequencias = ObterFrequenciaAula() }).ToList();
         }
 
-        private IEnumerable<FrequenciaAulaDto> ObterFrequenciaAula(string codigoAluno)
+        private IEnumerable<FrequenciaAulaDto> ObterFrequenciaAula()
         {
             return QUANTIDADES_AULAS.Select(numeroAula => new FrequenciaAulaDto
             {
                 NumeroAula = numeroAula,
                 TipoFrequencia = TIPOS_FREQUENCIAS[new Random().Next(TIPOS_FREQUENCIAS.Length)].ObterNomeCurto()
             }).ToList();
-        }        
+        } 
+        
+        protected IEnumerable<FrequenciaSalvarAlunoDto> ObterListaFrequenciaSalvarAlunoComAusencia()
+        {
+            return CODIGOS_ALUNOS.Select(codigoAluno => new FrequenciaSalvarAlunoDto
+                { CodigoAluno = codigoAluno, Frequencias = ObterFrequenciaAula(codigoAluno) }).ToList();
+        }
+
+        private IEnumerable<FrequenciaAulaDto> ObterFrequenciaAula(string codigoAluno)
+        {
+            string[] codigosAlunosAusencia = { CODIGO_ALUNO_1, CODIGO_ALUNO_3 };
+            string[] codigosAlunosPresenca = { CODIGO_ALUNO_2, CODIGO_ALUNO_4, CODIGO_ALUNO_6 };
+            string[] codigosAlunosRemotos = { CODIGO_ALUNO_5 };
+
+            return QUANTIDADES_AULAS.Select(numeroAula => new FrequenciaAulaDto
+            {
+                NumeroAula = numeroAula,
+                TipoFrequencia = codigosAlunosAusencia.Contains(codigoAluno) ? TipoFrequencia.F.ObterNomeCurto() :
+                    codigosAlunosPresenca.Contains(codigoAluno) ? TipoFrequencia.C.ObterNomeCurto() :
+                    codigosAlunosRemotos.Contains(codigoAluno) ? TipoFrequencia.R.ObterNomeCurto() :
+                    TIPOS_FREQUENCIAS[new Random().Next(TIPOS_FREQUENCIAS.Length)].ObterNomeCurto()
+            }).ToList();
+        }
 
         private async Task CriarAulas(long componenteCurricularId, int bimestre, bool turmaHistorica)
         {
